Remove too hard goal from roadmap.
EPERM you actually do want to watch them, because other users can
modify them. Figuring out if a file is on a read-only folder is
actually hard, and likely only a good thing to do on distributions
such as Fedora Silverblue, so not really the best thing to spend time
on.