Build a "static" lisp image.
It's not a static binary in the classical sense, but shared libraries
such as osicat's are at least bundled, which only leaves
libc/libz/libm/libdl as the shared libraries to use.

This does mean that suddenly, we need to care about glibc, hence
building for every distribution in the world. It's okay.